Fix handle[Long|Double]ValueOf failing to access symbolic operands
Longs and doubles take up 2 slots per variable. The symbolic
values of the corresponding variables are always associated with
the second slot of the long or double.

Because of this, we have to use getOperandAttr with an offset of 1
rather than 0 to get the symbolic values of the String.valueOf
parameters when called with a long or double argument.